cleans up commented out methods, adds customer repo helper module
