+++ lib/repository_helper_modules/helper.rb
@@ -1,7 +1,9 @@
 require_relative './invoice_helper.rb'
 require_relative './invoice_item_helper.rb'
+require_relative './customer_helper.rb'
 
-module Helper
+module RepositoryHelper
   include InvoiceHelper
   include InvoiceItemHelper
+  include CustomerHelper
 end
